rename dir for radicale proxy_controller
